Move persona finding from the base class to a fetcher class
