+++ lib/personas/persona_routing_constraint.rb
@@ -3,7 +3,7 @@
 module Personas
   class PersonaRoutingConstraint
     def initialize(persona_type)
-      @persona_type = RoutablePersonas.find(persona_type)
+      @persona_type = RoutablePersonas.find(persona_type.to_s)
     end
 
     def matches?(request)
